@@ -7,19 +7,19 @@
     "webextension-polyfill-ts": "0.22.0"
   },
   "devDependencies": {
-    "webpack": "5.3.2",
-    "webpack-cli": "4.1.0",
-    "find": "0.3.0",
-    "babel-jest": "26.6.3",
-    "jest": "26.6.3",
-    "ts-jest": "26.4.4",
+    "@babel/core": "7.16.0",
+    "@types/jasmine": "3.10.2",
+    "@types/jest": "27.0.3",
+    "babel-jest": "27.4.2",
+    "babel-loader": "8.2.3",
     "file-loader": "6.2.0",
-    "ts-loader": "8.0.7",
-    "babel-loader": "8.1.0",
-    "typescript": "4.0.5",
-    "@babel/core": "7.12.3",
-    "@types/jasmine": "3.6.0",
-    "@types/jest": "26.0.19"
+    "find": "0.3.0",
+    "jest": "27.4.2",
+    "ts-jest": "27.0.7",
+    "ts-loader": "9.2.6",
+    "typescript": "4.5.2",
+    "webpack": "5.64.4",
+    "webpack-cli": "4.9.1"
   },
   "jest": {
     "preset": "ts-jest"
