@@ -4,20 +4,20 @@
   "private": true,
   "homepage": "/options/",
   "dependencies": {
-    "@types/jest": "26.0.15",
-    "@types/node": "14.14.6",
-    "@types/react": "16.9.55",
-    "@types/react-dom": "16.9.9",
-    "@types/react-router-dom": "5.1.6",
-    "@types/material-ui": "0.21.8",
-    "@material-ui/core": "4.11.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.0",
+    "@material-ui/core": "4.12.3",
+    "@types/jest": "27.0.3",
+    "@types/material-ui": "0.21.12",
+    "@types/node": "16.11.11",
+    "@types/react": "17.0.37",
+    "@types/react-dom": "17.0.11",
+    "@types/react-router-dom": "5.3.2",
+    "react": "^17.0.2",
     "react-axe": "3.5.4",
-    "webextension-polyfill-ts": "0.22.0",
-    "typescript": "4.0.5"
+    "react-dom": "^17.0.2",
+    "react-router-dom": "6.0.2",
+    "react-scripts": "4.0.3",
+    "typescript": "4.5.2",
+    "webextension-polyfill-ts": "0.22.0"
   },
   "scripts": {
     "start": "react-scripts start",
